feat : step3 피드백  - 문자열 계산기 enum 계산기 객체를 도출
