+++ README.md
@@ -13,7 +13,7 @@
 * step3 피드백 반영
     * ~~4 미만의 값을 파라미터로 전달하면 자동차가 멈춰있는지에 대한 테스트 추가~~
     * ~~출력을 CarResultView 출력 객체 내부에서 수행~~
-    * 문자열 계산기 - enum or 각 연산에 맞는 계산기 객체를 도출
+    * ~~문자열 계산기 - enum or 각 연산에 맞는 계산기 객체를 도출~~
 * ~~1자 이상, 5자 이하의 정상적인 이름인지 확인~~
 * ~~참여자 이름 split한 후 자동차 생성~~
 * ~~경주에 참여한 자동차 중에서 우승자 찾기~~
